Add Abyssal Sire Plugin
Add Abyssal Sire Plugin for toggling left-click to walk under Spawns and Scions.

This plugin improves the QoL of the Abyssal Sire by deprioritizing the Attack option on the Spawns and Scions, allowing you to left-click to walk on them. This plugin solves a similar problem for Abyssal Sire that the "left click walk on core" plugin option does for Corporeal Beast.